pyproject.toml
[tool.poetry]
name = "retro-arcade"
version = "0.1.0"
description = ""
authors = ["Chris Proctor <chris@chrisproctor.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
retro-games = "^0.1.3"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
